Add layout system.

+++ src/DotX/Brush/IBrush.cs
@@ -4,6 +4,6 @@ namespace DotX.Brush
 {
     public interface IBrush
     {
-        void Render(Context context, int width, int height);
+        void Render(Context context, double width, double height);
     }
 }
